[Validators] (refactor) moved software_catalog validator to the validators package
